wallet dialogs renamed
