+++ app/components/wallet/add/WalletOptionsDialog.scss
